refactor: Migrate from InternalResolvedUnitResult
