+++ lib/src/obsoleted/anti_patterns/long_method.dart
@@ -1,10 +1,11 @@
+import 'package:analyzer/dart/analysis/results.dart';
+
 import '../../metrics/source_lines_of_code/source_code_visitor.dart';
 import '../../models/function_type.dart';
 import '../../models/issue.dart';
 import '../../models/scoped_function_declaration.dart';
 import '../../utils/metric_utils.dart';
 import '../constants.dart';
-import '../models/internal_resolved_unit_result.dart';
 import 'obsolete_pattern.dart';
 import 'pattern_utils.dart' as utils;
 
@@ -17,7 +18,7 @@ class LongMethod extends ObsoletePattern {
 
   @override
   Iterable<Issue> legacyCheck(
-    InternalResolvedUnitResult source,
+    ResolvedUnitResult source,
     Iterable<ScopedFunctionDeclaration> functions,
     Map<String, Object> metricsConfig,
   ) {
@@ -30,7 +31,7 @@ class LongMethod extends ObsoletePattern {
     final issues = <Issue>[];
 
     for (final function in functions) {
-      final visitor = SourceCodeVisitor(source.unit.lineInfo!);
+      final visitor = SourceCodeVisitor(source.unit!.lineInfo!);
       function.declaration.visitChildren(visitor);
 
       if (visitor.linesWithCode.length > threshold) {
