+++ lib/src/obsoleted/anti_patterns/pattern_utils.dart
@@ -1,22 +1,22 @@
 // ignore_for_file: long-parameter-list
+import 'package:analyzer/dart/analysis/results.dart';
 import 'package:analyzer/dart/ast/ast.dart';
 import 'package:source_span/source_span.dart';
 
 import '../../models/issue.dart';
 import '../../models/severity.dart';
-import '../models/internal_resolved_unit_result.dart';
 import 'obsolete_pattern.dart';
 
 Issue createIssue(
   ObsoletePattern pattern,
   String message,
   String recommendation,
-  InternalResolvedUnitResult source,
+  ResolvedUnitResult source,
   Declaration issueNode,
 ) {
-  final offsetLocation = source.unit.lineInfo
+  final offsetLocation = source.unit?.lineInfo
       ?.getLocation(issueNode.firstTokenAfterCommentAndMetadata.offset);
-  final endLocation = source.unit.lineInfo?.getLocation(issueNode.end);
+  final endLocation = source.unit?.lineInfo?.getLocation(issueNode.end);
 
   return Issue(
     ruleId: pattern.id,
@@ -34,7 +34,7 @@ Issue createIssue(
         line: endLocation?.lineNumber,
         column: endLocation?.columnNumber,
       ),
-      source.content.substring(issueNode.offset, issueNode.end),
+      source.content!.substring(issueNode.offset, issueNode.end),
     ),
     severity: Severity.none,
     message: message,
