+++ test/obsoleted/anti_patterns/long_parameter_list/examples/example.dart
@@ -1,0 +1,5 @@
+void main() {}
+
+void func(int a, int b, int c, int e) {}
+
+void func2(int a, int b, int c, int e, String f) {}
